requirements.txt
absl-py==0.7.1
albumentations==0.4.3
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.6.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==19.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imageio==2.6.1
imgaug==0.2.6
incremental==16.10.1
ipdb==0.12.2
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.4
notebook==6.0.0
numpy==1.16.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
opencv-python==4.1.2.30
PAM==0.4.2
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pbr==3.1.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.1.1
pyRFC3339==1.0
pyrsistent==0.15.3
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.16.2
scikit-learn==0.21.2
scipy==1.3.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
testresources==2.0.0
Theano==1.0.4
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.38.0
traceback2==1.4.0
traitlets==4.3.2
Twisted==17.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.0
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
